Add params to after_perform callback
This is making the interface of BaseContracted a bit
more consistent where params are passed to other callbacks
such as before_perform as well.

This change effectively allows to implement behaviour in
after_perform callbacks that's based on parameters without
using workarounds such as storing parameters in another callback
first.